requirements.txt
supabase==2.7.4
google-generativeai==0.7.2
python-dotenv==1.0.0
tiktoken==0.7.0
PyPDF2==3.0.1
python-docx==1.1.0